requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3

requests==2.28.1

pdfminer.six==20221105
python-docx==0.8.11
Pillow==9.3.0

# Optional for running Hugging Face models locally (not used in this app)
transformers==4.26.0
torch==1.13.1
